% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_apply.R
\name{as_Net}
\alias{as_Net}
\title{Constructor for single Net object}
\usage{
as_Net(matrix, net.name, node.variables)
}
\arguments{
\item{matrix}{Network represented as an adjacency matrix}

\item{net.name}{Name of the network (optional)}

\item{node.variables}{Node level variables (optional)}
}
\value{
A Net instance
}
\description{
This function takes a single network, as an adjacency matrix, and returns a
Net object.
}
