% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_apply.R
\name{group_diff_test}
\alias{group_diff_test}
\title{Group difference test}
\usage{
group_diff_test(netSampleStatSet, grouping.variable, p.adjust = "BY",
  non.parametric = F)
}
\arguments{
\item{netSampleStatSet}{Input NetSampleStatSet}

\item{grouping.variable}{character name of sample level grouping variable}

\item{p.adjust}{character string for requested multiple comparisons
adjustment. Defaults to Benjamani-Hochberg}

\item{non.parametric}{Logical. if true, test is performed using Wilcox test.
If false, Welch's t-test. Defaults to false.}
}
\value{
A data frame containing original and adjusted p.values.
}
\description{
This function implements the group difference test on a network statistic.
This test assesses if the change in the network statistic due to the network
manipulation is significantly different between groups.
}
\details{
If the sample has 2 groups, this test is performed using Welch's t-test or
Wilcox test. If the sample has 3 or more groups, the test is performed using
a 1-way ANOVA, or Kruskal-Wallis test. Differences are tested at each network
manipulation.
}
