% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GroupA}
\alias{GroupA}
\title{Simulated Dataset of 20 networks. Group A.}
\format{A `list` of 20 `matrix` objects, representing adjacency matrices. Node 10 is simulated to be important for global efficiency}
\usage{
GroupA
}
\description{
A simulated dataset of 20 binary networks with 20 nodes each.
}
\keyword{datasets}
