% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GroupB}
\alias{GroupB}
\title{Simulated Dataset of 20 networks. Group B.}
\format{A `list` of 20 `matrix` objects, representing adjacency matrices.}
\usage{
GroupB
}
\description{
A simulated dataset of 20 binary networks with 20 nodes each. Node 15 is simulated to be important for global efficiency.
}
\keyword{datasets}
