% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Net-class}
\alias{Net-class}
\title{An S4 class to represent a single network.}
\description{
This class represents a single observation of a network, with associated node level variables.
}
\details{
For constructor see: \code{\link{as_Net}}
}
\section{Slots}{

\describe{
\item{\code{net}}{matrix. The network represented as an adjacency matrix.}

\item{\code{net.name}}{character. The name of the network (e.g. subject ID, school name)}

\item{\code{node.variables}}{list. A named list of node variables, in the same order as the adjacency matrix.}
}}

