% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_apply.R
\docType{methods}
\name{net_stat_apply}
\alias{net_stat_apply}
\alias{net_stat_apply,NetSet-method}
\alias{net_stat_apply,NetSampleSet-method}
\title{Apply a network statistic function to a NetSet or NetSampleSet object.}
\usage{
net_stat_apply(netSet, net.stat.fun, net.stat.fun.args, net.stat.name)

\S4method{net_stat_apply}{NetSet}(netSet, net.stat.fun, net.stat.fun.args,
  net.stat.name)

\S4method{net_stat_apply}{NetSampleSet}(netSet, net.stat.fun, net.stat.fun.args,
  net.stat.name)
}
\arguments{
\item{netSet}{A NetSet or NetSampleSet object.}

\item{net.stat.fun}{The network statistic function}

\item{net.stat.fun.args}{A list of additional arguments to the network
statistic function}

\item{net.stat.name}{A descriptive name for the network statistic (defaults to
deparsed name of statistic function)}
}
\value{
A NetStatSet or NetSampleStatSet
}
\description{
This function applies a network statistic function to a NetSet or NetSampleSet
object, and returns the calculated network statistics.
}
\section{Methods (by class)}{
\itemize{
\item \code{NetSet}: net_stat_apply for NetSet

\item \code{NetSampleSet}: Converter for NetSampleSet
}}

