% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NetSample-class}
\alias{NetSample-class}
\title{An S4 class to represent a sample of networks}
\description{
This class represents a collection of networks, with associated network level variables.
}
\details{
For constructor see: \code{\link{as_NetSample}}
}
\section{Slots}{

\describe{
\item{\code{nets}}{list. A list of Net objects}

\item{\code{net.names}}{character. A character vector representing the names of the net objects}

\item{\code{sample.variables}}{list. A named list of network level variables.}
}}

