% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NetSampleSet-class}
\alias{NetSampleSet-class}
\title{An S4 class representing a sample of networks with a network permutation function applied to it.}
\description{
This class represents the results of applying a network permutation function, such as a jackknife, or rewiring algorithm, to a sample of networks.
}
\section{Slots}{

\describe{
\item{\code{net.sets}}{list. A list of NetSet objects, reach representing a network, and the results of applying the permutation function.}

\item{\code{net.names}}{character. A character vector representing the names of the original network in the sample}

\item{\code{sample.variables}}{list. A list representing sample level variables.}
}}

