% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NetStatSet-class}
\alias{NetStatSet-class}
\title{An S4 class representing the results of applying a network statistic function to a single NetSet object.}
\description{
This class represents the results of applying a network statistic function to a single NetSet object. This class contains the results for the original network, as well as for each instance of the permuted/manipulated networks.
}
\section{Slots}{

\describe{
\item{\code{stat.fun}}{function. The network statistic function applied}

\item{\code{stat.fun.name}}{character. The name of the network statistic function}

\item{\code{stat.fun.args}}{list. Additional arguments the network statistic function took}

\item{\code{orig.net.name}}{character. The name of the original network.}

\item{\code{orig.net.stat}}{numeric. The value of the network statistic calculated on the original network.}

\item{\code{nets.stat}}{list. A list of values of the network statistic applied to the manipulated networks}

\item{\code{nets.names}}{character. Names of the manipulated networks.}
}}

