% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{network_functions}
\alias{network_functions}
\alias{node_jackknife}
\alias{network_jackknife}
\alias{absolute_threshold}
\alias{relative_threshold}
\title{Network Manipulation Functions}
\usage{
node_jackknife(Net)

network_jackknife(Net, network.variable)

absolute_threshold(Net, thresholds)

relative_threshold(Net, percentiles)
}
\arguments{
\item{Net}{Network to jackknife}

\item{network.variable}{character name of node variable containing network labels}

\item{thresholds}{Vector of thresholds to use}

\item{percentiles}{Vector of densities to threshold at}
}
\value{
A list of Net objects

A list of Net objects

A list of Net objects, named at threshold level.

List of net objects with named at the density threshold.
}
\description{
Network Manipulation Functions

Removes each node in turn from the Net, and returns a list of jackknifed Net objects

Removes each subnetwork in turn from the Net, and returns a list of jackknifed Net objects

Applies an absolute thresholding to target network, returning a binary network.

Applies a relative threshold based on target density to target network
}
