% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{pretty_print_arg_list}
\alias{pretty_print_arg_list}
\title{Format and print an argument list.}
\usage{
pretty_print_arg_list(arg.list)
}
\arguments{
\item{arg.list}{list of arguments to format.}
}
\value{
Formatted argument list
}
\description{
Internal function for pretty printing argument lists.
}
\keyword{internal}
