% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{show,Net-method}
\alias{show,Net-method}
\alias{show,NetSample-method}
\alias{show,NetSet-method}
\alias{show,NetSampleSet-method}
\alias{names,NetSet-method}
\alias{show,NetStatSet-method}
\alias{show,NetSampleStatSet-method}
\alias{show-Net}
\title{Convenience Functions for printing information about Net-type objects}
\usage{
\S4method{show}{Net}(object)

\S4method{show}{NetSample}(object)

\S4method{show}{NetSet}(object)

\S4method{show}{NetSampleSet}(object)

\S4method{names}{NetSet}(x)

\S4method{show}{NetStatSet}(object)

\S4method{show}{NetSampleStatSet}(object)
}
\arguments{
\item{object, x}{Object to print}
}
\description{
Convenience Functions for printing information about Net-type objects
}
