% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_apply.R
\name{diff_test}
\alias{diff_test}
\title{Test for differences from original statistic}
\usage{
diff_test(netSampleStatSet, p.adjust = "BH", non.parametric = F)
}
\arguments{
\item{netSampleStatSet}{Input NetSampleStatSet}

\item{p.adjust}{character string for requested multiple comparisons
adjustment. Defaults to Benjamani-Hochberg}

\item{non.parametric}{Logical. if true, test is performed using Wilcox test.
If false, Welch's t-test. Defaults to false.}
}
\value{
A data frame containing original and adjusted p.values, as well as
  differences, labeled with manipulation name.
}
\description{
This function tests for significant differences from the original network
statistic as a result of the network manipulation. If non-parametric is
chosen, this is done using the Wilcox test, otherwise, Welch's t-test.
}
