% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_statistics.R
\name{network_statistics}
\alias{network_statistics}
\alias{global_efficiency}
\alias{modularity}
\title{Network Manipulation Functions}
\usage{
global_efficiency(Net)

modularity(Net, community.variable)
}
\arguments{
\item{Net}{Input Net}

\item{community.variable}{character name of the node variable that represents the partition.}
}
\value{
global efficiency value

modularity value
}
\description{
Network Manipulation Functions

Computes global efficiency

Computes modularity with a given partition
}
